fix master

+++ backend/src/main/java/com/example/backend/master/platform/provisioningLogs/ProvisioningLogsRepository.java
@@ -2,5 +2,5 @@ package com.example.backend.master.platform.provisioningLogs;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ProvisioningLogsRepository extends JpaRepository<ProvisioningLogs, Integer> {
+public interface ProvisioningLogsRepository extends JpaRepository<ProvisioningLogs, Long> {
 }
