+++ backend/src/main/java/com/example/backend/master/platform/systemUsers/SystemUsersRepository.java
@@ -2,5 +2,5 @@ package com.example.backend.master.platform.systemUsers;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SystemUsersRepository extends JpaRepository<SystemUsers, Integer> {
+public interface SystemUsersRepository extends JpaRepository<SystemUsers, Long> {
 }
