+++ backend/src/main/java/com/example/backend/master/platform/tenantAdminUsers/TenantAdminUsersRepository.java
@@ -2,5 +2,5 @@ package com.example.backend.master.platform.tenantAdminUsers;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TenantAdminUsersRepository extends JpaRepository<TenantAdminUsers, Integer> {
+public interface TenantAdminUsersRepository extends JpaRepository<TenantAdminUsers, Long> {
 }
