+++ backend/src/main/java/com/example/backend/master/platform/tenantDatabases/TenantDatabasesRepository.java
@@ -2,5 +2,5 @@ package com.example.backend.master.platform.tenantDatabases;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TenantDatabasesRepository extends JpaRepository<TenantDatabases, Integer> {
+public interface TenantDatabasesRepository extends JpaRepository<TenantDatabases, Long> {
 }
