+++ backend/src/main/java/com/example/backend/master/platform/tenants/TenantsRepository.java
@@ -2,5 +2,5 @@ package com.example.backend.master.platform.tenants;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TenantsRepository extends JpaRepository<Tenants,Integer> {
+public interface TenantsRepository extends JpaRepository<Tenants, Long> {
 }
